Add colibri-ut library - a common unit test library
The intent of this library is to include all code, which could be potentially
useful for several UTs and thus can be shared, avoiding duplication of
similar code.

This library is linked as colibri module and not as colibri -ut library. This
allows to use it in both - UTs and in standalone programs which don't use CUnit.

At the moment of creation, it contains only RPC helper functions:

    c2_rpc_client_(init|fini)
    c2_rpc_server_(init|fini)

which are moved from rpc/rpclib.c