configure.ac: add support for colibri-cunit repository
With this commit configure expects a CUnit sources to be present in
../cunit/ directory. CUnit is built while running configure, the same
way as it's done for db4, galois and yaml. System CUnit library is not
required any more.

Two new variables are available for substitution in Makefile.{am,in} if
some component requires to refer libcunit.a or CUnit include path:

CUNIT_LIBS
CUNIT_HEADERS

Tested-by: Dmitriy Chumak <dmitriy_chumak@xyratex.com>