rpcping: report correct statistics on client side
In order to get a correct statistics from rpc machine, we should extract
it at the moment when all connections are already terminated, but rpc
machine is not finalized yet. It's not possible to do with
c2_rpc_client_fini().

So the rpcping-specific client_fini() function, which is introduced in
this commit, is used instead of c2_rpc_client_fini().