C2_TRACE_POINT() macro code size optimization
From Nikita:
C2_TRACE_POINT() is going to be used in many places. The amount of
C code (which means the amount of machine code instructions) this
macro is expanded into if very important, because even the slight
increase in it would increase the size of Colibri binary image
considerably thus increasing our instruction cache foot-print.