lib: add basic fault injection API
Fault Injection API provides functions to set "fault points" inside the
code, and functions to enable/disable the failure of those points. It's
aimed at increasing code coverage by enabling execution of
error-handling code paths, which are not covered otherwise by unit
tests.

The API is enabled only when ENABLE_FAULT_INJECTION macro is defined.
Otherwise it's automatically compiled out to avoid unneeded overhead in
production code. ENABLE_FAULT_INJECTION is enabled by default, but it
can be disabled with --disable-finject option of configure script.

A basic API consists of the following functions and macros:

    int c2_fi_init(void);
    void c2_fi_fini(void);

    bool C2_FI_ENABLED(tag)

    void c2_fi_enable(func, tag)
    void c2_fi_enable_once(func, tag)
    void c2_fi_enable_random(func, tag, p)
    void c2_fi_enable_off_n_on_m(func, tag, n, m)
    void c2_fi_enable_each_nth_time(func, tag, n)
    void c2_fi_enable_func(func, tag, trigger_func, data)

    void c2_fi_disable(func, tag);

A typical use case is to place a fault point inside the function of
interest and do some action if fault point is enabled:

    void *c2_alloc(size_t n)
    {
     ...
     if (C2_FI_ENABLED("pretend_failure"))
        return NULL;
     ...
    }

Fault point is defined in place of C2_FI_ENABLED() usage. In this
example a fault point with tag "pretend_failure" is created in function
"c2_alloc", which can be enabled/disabled from external code with
something like the following:

    c2_fi_enable_once("c2_alloc", "pretend_failure");

Reviewed-on: http://reviewboard.clusterstor.com/r/732