fop/fom: c2_fom_callback_fini() is optional now
Rationale: code safety (user may forget to call fini).
The actual fini is performed in the AST callback after calling
the bottom half now. The original idea of c2_fom_callback_fini()
was just to make sure that the callback can be freed and
there is no reason to force user to check this.