configure.ac: add checks for dependencies of gccxml2xcode
gccxml2code depends on perl and XML::LibXML perl module. Also it works
with an output of gccxml, which means that gccxml should be installed
either.

So configure will stop with error if perl, perl modules, gccxml or g++
are not available in current environment not found (g++ is required by
gccxml).

Reviewed-on: http://reviewboard.clusterstor.com/r/921