lib/trace: fix poor name of internal variable in C2_LOG macro
When choosing names for internal variables in macros, one should bear in
mind that they can clash with names of variables provided as parameters
to macro. So one should better use a longer name, prefixed with series
of underscores to reduce a possibility of name conflicts, which manifest
themselves in cryptic and non-obvious compilation errors.