Replace print[fk] with C2_LOG() in core modules
C2_LOG() is preferred over print/printk in source files, which
constitute libcolibri.so and kcolibri.ko (see doc/logging-and-tracing
for explanation of "why").

But on the other hand we don't want to replace printf/printk with C2_LOG
in such parts which deals with formatted user-interface-oriented output
like UT, ST, various utils, so they kept unchanged.