xcode: change type of c2_void_t to make it work in C++ as well
Consider the following C code:

    typedef struct {;} c2_void_t;

    struct fixarray {
            c2_void_t fa_none C2_XCA_TAG("NR");
            struct optfid *fa_data;
    } C2_XCA_SEQUENCE;

gcc will report a 0 offset for both fa_none and fa_data fields, but g++ doesn't.
g++ will report a 0 offset for fa_none and 8 offset for fa_data field.

Since we're using gccxml to generate xcode, and gccxml essentially is a C++
compiler, we need this code to produce the same 0 offsets in both C and
C++. That's why we need to change type of c2_void_t to be:

    typedef char c2_void_t[0];

Which will give a 0 offset for both fields fo fixarray struct from the
above example in C and C++.