xcode/gccxml2xcode: use a better name for a placeholder of an undefined xcode type
Normally, a 'void' C type can't be a valid type of a variable and thus
it doesn't have a corresponding xcode C2_XT_XXX type. But just in case
of some internal logic error in gccxml2xcode, when 'void' could "leak"
into generated code, we set xc_type for 'void' to 'UNDEFINED', so it
would be easier to sort such error out.