lib/trace: add CALL/DEBUG/INFO/WARN/ERROR/FATAL levels to C2_LOG()
Colibri trace levels are different from a "traditional" log-levels concept,
which for example is used in kernel's printk().

In Colibri, each trace level can be enabled/disabled separately, because they
are implemented as a bitmask. This is done for greater flexibility. But there
are special trace levels with '+' at the end of a level name, which emulate
"traditional" log-levels.

The syntax to specify a desired trace level is as following:

    level[+][,level[+]] where level is one of call|debug|info|warn|error|fatal

For example

    debug       -- enables only log messages with C2_DEBUG level
    info+       -- enables log messages with C2_INFO level and higher
    warn+       -- only C2_WARN, C2_ERROR and C2_FATAL messages are enabled,
                   this is the default trace level
    call+       -- all log messages are enabled
    call,error+ -- only C2_CALL, C2_ERROR and C2_FATAL are enabled

    call,debug,fatal -- only C2_CALL, C2_DEBUG and C2_FATAL

Notice, that trace messages with level C2_WARN and higher are printed
independently of trace immediate mask value. It means that they are always
printed, even if trace immediate mask is disabled for the subsystem to which
they belong. The only way to suppress C2_WARN, C2_ERROR or C2_FATAL messages is
disable them in trace level mask.

There are different ways to control trace level in user- and kernel- space.

User space:

    1) environment variable:

         export C2_TRACE_LEVEL=debug
         ./utils/ut.sh

    2) CLI options for utils/ut:

         -e     string: trace level: level[+][,level[+]] where level is one of
	                call|debug|info|warn|error|fatal

         $ ./utils/ut.sh -e call,warn+

       CLI option overrides environment variable

Kernel space:

    $ insmod kcolibri.ko trace_level='debug+'

Reviewed-by: Nikita Danilov <nikita_danilov@xyratex.com>
Reviewed-on: http://reviewboard.clusterstor.com/r/1015