console: fix link-time dependency on 'verbose' variable
It's completely wrong to include console/console.c into libcolibri.so
code base. But due to the fact that console module is written in a
sloppy manner it's not possible to do such separation in a right way w/o
some code reorganization and restructuring. So this 'verbose' variable
is added just to suppress linker error, and it may not work in the way
it should because console.c has it's own global variable of the same
name. This should be fixed someday. General guidelines is to keep only
necessary parts of console in libcolibri, everything else should be
factored out into c2console binary.