Add libcolibri-altogether to build targets
The idea behind "libcolibri-altogether" is pretty simple - to #include
all our .c files into a single .c file and then compile it into
libcolibri-altogether.so. This should provide more opportunities for
optimization to compiler, because all code is in a single translation
unit. A compiler can optimize much better within a compilation unit than
across them. If a file where all but one functions are static is
compiled, it is possible to completely eliminate all function calls and
their associated overhead (stack manipulations, jumps, pipeline
disruptions, &c.).

This patch adds new build target libcolibri-altogether.so and changes
colibri code so that it can be built as a single .c file:

 * undefine multiple-definition macros at the end of .c files (like
   C2_TRACE_SUBSYSTEM, etc.);

 * rename static variables and functions, which have identical names but
   defined in different .c files to avoid name clashes;

 * prevent redefinition of system macros like _POSIX_C_SOURCE or
   _ISOC99_SOURCE.