scripts/c2mount.sh: startup and mount Colibri fs
c2mount.sh script allows automatically startup the Colibri services
on arbitrary number of remote nodes and eventually mount the client
on the local node with just a one command. AD-STOBs are supported with
automatic generation of disc configuration files for the services
on Titan controllers.

Carl & Dave are initial authors of the script.
https://docs.google.com/a/xyratex.com/document/d/1kCZiRWgqBSlnTfcQZ6_REz0JQEXEWyFePoUJp8QSXbQ/view

See detailed documentation inside the script.