db/db: hack around db5 locking problems.
db4 uses commit-duration locks on pages, which results in deadlocks in Colibri
services. Work around this by disabling locking sub-system in db4 entirely and
using per-table mutex for synchronisation:

    * db/db.c:dbenv_setup(): remove DB_INIT_LOCK from the flags, so that locking
      sub-system is not initialised;

    * db/db_impl.h:c2_table_impl: add t_lock mutex serialising all operations on
      the table;

    * initialise and finalise the mutex in c2_table_{init,fini}();

    * db/db.c:{TABLE,CURSOR}_CALL(): use table mutex to serialise the operations
      against the table;

    * c2_db_cursor_init(): ignore all lock related cursor initialisation flags,
      because they lead to EINVAL errors with locking sub-system disabled;

    * db/ut/db.c: eliminate "cursor_flag_rmw" unit test.

This solution leaves much to be desired. For one thing, the mutex can be easily
replaced with a read-write lock, improving concurrency. However, being what it
is, that is, a short-term hack used until the proper solution (RVM) is in place,
I prefer to leave it as simple as possible.

Tests: UT (user, kernel), ST.