scripts/c2: allow 'dist-check' to work when colibri sources are on NFS
If colibri sources are stored on NFS filesystem, `make dist` command
will fail, when run as root, because it tries to preserve file's
ownership, which is not always allowed for 'root' over NFS. To address
this, we create dist archive in /tmp and unpack it w/o preserving
ownership.

Also, two environment variables, CONFIGURE_OPTS and MAKE_OPTS, were
added to pass custom options for `configure` and `make` commands.

Signed-off-by: Valery V. Vorotyntsev <valery_vorotyntsev@xyratex.com>