scripts/check-public-api: don't depend on current locale settings
check-public-api script could report fake API changes, like:

    Checking colibri public API...   FAIL
    WARNING: Did you forget to add the following public functions
             to the 'colibri/colibri-pub.api' list or to add C2_INTERNAL
             prefix to them?

    -c2_fini
    +c2_fini

It's because sorting order depends on a current locale settings (kudos
to Valery for this hint). To prevent such accidents, we explicitly set
"C" locale for `sort` command and do sort colibri-pub.api "on the fly"
to not depend on it's current sort order, because it can be messed by
manual editing.

Reported-by: Andriy Tkachuk <andriy_tkachuk@xyratex.com>
Tested-by:   Andriy Tkachuk <andriy_tkachuk@xyratex.com>