Optionally attach debugger to colibri executable on failure.
If COLIBRI_DEBUGGER environment variable is set when c2_init() is called,
c2_panic() would fork a process and execute specified debugger in the created
child process. The parent spins in an infinite loop, waiting for the debugger to
attach. The debugger must accept gdb command line parameters: "executable_name
pid". Note that environment variables do not survive su or sudo, so ./script/c2
cannot be used as is.

Reviewed-on: http://reviewboard.clusterstor.com/r/1110/
Reviewed-by: Dmitriy Chumak <dmitriy_chumak@xyratex.com>