rpc: Wake-up rpc-worker thread after each 60 seconds.
If thread doesn't change its state then kernels thinks the thread is
perhaps hung and prints its stack trace in system log.
In our case, rpc-worker is a thread that processes ASTs posted on
c2_rpc_machine::rm_sm_grp. rpc-worker can keep on waiting for ASTs
on a clink for a long time. This floods system log with stack traces
of rpc-worker thread.

To fix it just made an arrangement so that the thread wakes up
after every 60 seconds.