rpc: add ->rio_deliver() operation.
Add m0_rpc_item_ops::rio_deliver() operation called by the rpc layer on receiver
side to deliver the item to the upper layer.

Default implementation submits the item to the request handler.

This method is needed for items that are not following normal fom-based model of
processing (e.g., HA items).

<size>
    text     data      bss    total filename
     -66       32        0      -34 libmero.so
     -41        0        0      -41 m0mero.ko
    -186        0        0     -186 libmero-altogether.so
</size>
Reviewed-on: https://reviewboard.clusterstor.com/r/1134/
Reviewed-by: Amit Jambure <amit_jambure@xyratex.com>