Makefile.am: ensure that all-local target executed last during parallel build (-j)
Automake generates the following code for the main build target am-all:

    all-am: Makefile $(LTLIBRARIES) $(PROGRAMS) $(HEADERS) config.h all-local

For non-parallel builds, it means that all-local target will be
executed last. But for parallel (make -j) builds this is not the case.
Make can execute all-local target before other targets completes. To
prevent this we add our libraries and programs as prerequisites for
all-local target, to ensure that it will be executed only after other
targets are completed.

Also, the code to print sizes of elf images is factored out into a
separate script print-elf-sizes.