m0_fol_init() is changed to supports fol recoed parts based design.
m0_dtx is used instead m0_db_tx in m0_fol_init().
rec_fini() is changed to m0_rec_fini() which needs to be called
whenever FOL record is fetched and it's usage is over.

In FOL ut m0_xcode_cmp() is used to rec header of fetched record
and test record instead of memcmp().

m0_fol_rec_part_add() is added to add the FOL record part in the
record during updates.