configure.ac: rename option --with-linux-obj to --with-linux
This option is used not only to point where kernel build tree is
located, but also to specify the installation directory of kernel
modules. It's better to have a general option names first and then more
specific ones, if needed. For example, Lustre project uses a similar
approach. Also, dropping '-obj' part makes it more consistent with
other options like this (--with-lustre).