build: add support for lustre-devel rpm
This patch changes behavior of the --with-lustre option of configure
script.

It enables additional checks in configure, which ensure that lustre
directory contains Lustre's config.h and Module.symvers files, which are
required for a correct build of Mero. On practice this means that in
order to build Mer, either lustre-devel rpm packages should be installed
systemwide (in which case configure will pick it up automatically) or
Lustre should be built from sources and path to the Lustre's build
directory should be pointed to configure using --with-lustre option.

The static lustre_config.h and Lustre.symvers files, from
build_kernel_modules/ directory, are removed. Instead of them, confiugre
will create a symlink for lustre_config.h and substitute the full path
to Lustre's Module.symvers as @LUSTRE_SYMVERS@.