Removes common symbols from mero.
Common symbols are avoided because they break the compilation
toolchain that Parallel Scientific is currently using.

The -fno-common flag is used with gcc so common symbols are
avoided, and the errors in the code arising from the use of
such flag have been fixed mainly by fiddling with the scope
specifiers in declarations.

Reviewed-on: http://rb.clusterstor.com/r/1260/
Reviewed-by: Nikita Danilov <nikita_danilov@xyratex.com>