scripts/m0: update 'dist-check' command
It seems to be a bug in automake - it doesn't handle 'distdir' parameter
correctly for subpackages (in our case extra-libs/galois). So we revert
back to the old behavior of `m0 dist-check` command and create dist
archive in the mero source tree instead of /tmp.

Also, galois is included into dist archive now, so there is no need to
create a symlink for it.