rpc: Allow m0_reqh_fop_handle() to return error code.
Without this patch, when fom allocation failed in m0_reqh_fop_handle()
there was no way to report it to RPC. RPC assumed that it had delivered
a request item and waiting for reply. In fom allocation failure case,
there is no reply. As a result, the request item was forever left in
IN_PROGRESS stage and session in BUSY stage, leading to failure in
cleanup stages.