mgmt,reqh: start services asynchronously using a FOM
Provided support for asynchronous start of request handler services
using a FOM hosted by the management service.  Please also see
https://docs.google.com/a/xyratex.com/document/d/10VtuJSH8gcMNjaS7wEvgDbn2v1vS7m2Czgi8nb857sQ/edit#heading=h.dxeg2o2lqyuf
for background information.

This FOM was originally planned to be part of a management "Service
RUN" FOP; I only implemented the FOM and not the FOP though the
naming and pattern used in the code matches existing FOPs in the
management code, and supports the future implementation of the
service run FOP.

I preserved the existing semantics of the mero/setup.c logic but
internally services start asynchronously and failed services are
left on the request handler list for later cleanup.  Service
developers can provide the alternate rso_start_async() method to
fully take advantage of the new functionality; the rso_start()
method is still supported.  I did not retrofit any existing services
to use the new method - a working example of the method can be seen
in the UT.

The best way to understand the new code is to start with the changes
in the state diagram in reqh/reqh_service.h (best viewed in RB) to
understand the new rso_start_async() method and its
m0_reqh_service_starting()/m0_reqh_service_started()/m0_reqh_service_failed()
support apis.

Reviewed-on: http://reviewboard.clusterstor.com/r/1369
Reviewed-by: Andriy Tkachuk <andriy_tkachuk@xyratex.com>
Reviewed-by: Mandar Sawant <mandar_sawant@xyratex.com>

Bloated-by:
<size>
    text     data      bss    total filename
   10598     3016       64    13678 libmero.so
    3000       64        0     3064 mero/m0mero.ko
</size>

Squash-merged-from: 1166f92 (mgmt-startup branch).