fop/fom: m0_fom_callback_arm(): use oneshot clink
cb_done() was taking the channel lock to remove
clink from the channel. This did not allow us
to arm the callbacks on the channels protected
by the same locality lock. Now, with oneshot flag
clink is removed automatically from the channel.

From another side, m0_fom_callback_cancel() requires
the channel to be protected by the user from now.