ut: change default behavior of M0_UT_ASSERT() to panic in kernel mode
Since our UT are written in somewhat incorrect manner, i.e. w/o error
checking and state recovery, it doesn't make sense to allow tests to
continue after a failed M0_UT_ASSERT(), because this would trigger
further errors, which eventually lead to a crash. And in this case it
would be hard to identify the original failed M0_UT_ASSERT() instance.

Such behavior is already implemented for user-space version of
M0_UT_ASSERT() and this patch enables similar behavior for the
kernel-space version.