fom_generic: tx_ prefix renamed to fom_tx_
This is correction to the previous commit 674e4b1:
"fom_generic: loc_ctx_ prefix renamed to tx_".
There was compilation conflict error with be/seg_dict
when altogether build mode was enabled (by default).
Unfortunately, in my build --disable-altogether-mode
configuration option was enabled - that's why
this dummy error was missed.