ioservice/io_foms: allocate credit ivec at fom_create()
Generic fom M0_FOPH_TXN_OPEN phase, where we
calculate the credits, is not designed to return
the error. But we need to allocate the indexvec
of all IO segments to pass it to m0_stob_write_credit().

So now we allocate this array at m0_io_fom_cob_rw::fcrw_ivec
in m0_io_fom_cob_rw_create() and return error if
allocation failed.

Note: we still free that indexvec at the same
stob_write_credit() routine when we finished to
calculate the credit and the indexvec is no longer needed.

Suggested-by: Madhavrao Vemuri <madhav_vemuri@xyratex.com>
References: http://reviewboard.clusterstor.com/r/1552