Add "wait" debugger.
Add a new type of "debugger": if MERO_DEBUGGER is set to "wait",
m0_debugger_invoke() causes a user space mero process to print its pid and sleep
indefinitely waiting for a debugger to attach.

This is useful when usual debugger attaching is not possible, e.g., when
multiple processes are started form scripts (perhaps on a different node), and
debuggers would end up fighting for console access.