MERO-62: lib/trace: loss of messages in case of m0d panic
Fix MERO-62 issue, it was due to a bug in m0tracedump utility, which was
crashing and failing to decode all trace records from a trace file.
There were no loss of trace records, all of them were captured in a
trace file.

The bug was in escape_yaml_str() func, which incorrectly calculated the
size of a remaining part of the string, so memmove() used to overwrite
'\0' at the end of string.