stob/linux_adieu.c: expensive check at ioq_queue_get/put()
The assert check for the queue length at ioq_queue_get/put()
routines is set as expensive (by the profiling results).
Indeed, m0_queue_length() traverse all the queue to get
the length value and ioq_queue_get/put() routines are called
in the loop by themself, so that is huge overhead.