m0mount.sh: eliminate dependency on fdisk's output format
In order to find unused block devises, m0mount.sh exploited certain
keywords in the output of `fdisk` utility, in particular it relied on
this text line:

  Disk /dev/XXX doesn't contain a valid partition table

Unfortunately, newer fdisk versions don't print it any more. So, instead
of relying on a brittle fdisk's output format, this patch provides more
portable and consistent way to check whether a block device has a
partition table or doesn't, based on `partprobe` utility.