scripts/m0: don't leak disk space
We want to use the same value of $RUNDIR. If the values are different,
the script will not delete directory with old output and such directories
will proliferate.

When `m0' script is executed, it clears $RUNDIR directory and then populates
it with new output. Abstain from re-executing the script if you value this
output.