lib/timer: hard timer can be bound to an entire process
It was a problem with M0_TIMER_HARD timers.
If hard timer is attached to a timer locality, then it is bound to
some thread from timer locality. If thread finishes before timer
expires then timer expiration signal is lost and timer callback
is never called.

This patch fixes this problem by adding possibility to send signal to
a process and not to a thread if hard timer isn't attached to any
locality.

This patch also includes a lot of timer-related code refactoring.

Reviewed-on: https://reviewboard.clusterstor.com/r/1592/
Reviewed-by: Nikita Danilov <nikita_danilov@xyratex.com>
Reviewed-by: Yuriy Umanets <yuriy_umanets@xyratex.com>