addb_ut_mon_infra: fix UT race.
addb mon infra UT creates a sequence of addb records, which trigger addb
monitors, which send fops to the stats service. The UT then checks that the
service processed all the fops. The synchronisation between UT and stats service
was a simple m0_nanosleep(), which is unreliable. This led to sporadic failures.

I suspect (and shall assume until proven wrong) that this was the reason behind
https://jira.xyratex.com/browse/MERO-119 and
https://jira.xyratex.com/browse/MERO-78.

Fixed by patching stats fom finaliser (->fo_fini()) and upping a UT-visible
semaphore there.

In addition, fixed a number of minor stylistic issues.