The layouts per unit size are created at mds now
Today, unit_size is specified as mount parameter to
the whole m0t1fs.  It was temporary simplification.
Now, when confd enters onto the scene and we are
getting rid of specifying m0t1fs parameters via mount
command, it is the time to implement unit_size
parameter per each file which will be maintained by
layout module.

Some default unit_size should be provided by confd
(or, later, as extended attribute of correspondent
directory where the file is created).  Specific
unit_size for the file could be assigned to it via
setfattr command just after creation with touch
command. Clovis interface, which is aimed to be our
primary fs interface instead of POSIX, will allow
to specify unit_size of the file (as well as any
other attribute) along with it at the time of creation.
If the file already contain some data, its unit_size
can not be changed.

Reviewed-by: Nikita Danilov <nikita_danilov@xyratex.com>
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/2223
References: https://jira.xyratex.com/browse/MERO-114

Bloated-by:
<size>
    funcs    text     data      bss    total filename
       0        0        0        0        0
    3523     9425     3512       32    12969 libmero.so
   -1360    -4071        0        0    -4071 m0mero.ko
    2259    12624     4128       32    16784 libmero-altogether.so
</size>

Change-Id: I2a62b3c27ff172d03bdd13c0177f8810dad26525