git-submodules: update submodule URLs after moving to Gerrit
Update submodule URLs to anonymous gerrit url for each submodule
repository. This will provide a read-only access by default for
submodules. If one needs to have a write (commit) access, he should
change the "remote.origin.url" in submodule's repo to his ssh link.
For example:

  cd extra-libs/cunit/
  git config remote.origin.url ssh://dchumak@es-gerrit.xyus.xyratex.com:29418/mero-cunit

Note, that it requires to specify a username of a particular gerrit
user - "dchumak" in the above example.