dtm-ut: fix a race in dtm-transmit-ut:redo test.
Original code was witing until the reqh is idle. But it is possible that reqh is
idle because none of foms started yet, rather than because all of them
terminated.

Make sure that the foms have started by waiting on a special semaphore.

Reported-by: Nicholas Clarke <Nicholas_Clarke@xyratex.com>
Reported-at: https://jira.xyratex.com/browse/MERO-242

Change-Id: I6487fc32e3b218b0415cb8c813c46f73a69ece4d