MERO-244: Drop rpcs coming out of order to server.
Some rpc items now have xid assigned. This xid is passed to server.
Server keeps its own xid counter. If an item comes to server
with xid that violates ordering then the item is dropped same
way as it was dropped by "drop_item" fault injection.

This should fix resending of timed out rpc which was already handled
by server and its buffers are unmapped.

Note, that it is possible that rpcs may come out of order in normal
case (no timeout) as networking layer does not guarantee the correct
order. This is not the problem because such a dropped rpc will cause
client to resend it once again in correct order. Given this is quite
rare case it is ok to have this behavior.

Change-Id: I368d259d68858336b1060c3a0b4f71496f8b2114