rpc: reply fop usage simplification
There was number of m0_fop_put()s for the reply fops
throughout the code after m0_fom_init() just for no
particular clear reason. This way we just tried to
balance the reference counter of reply fops. But if
we take the reference not at the m0_rpc_item_send()
which is used for sending the replies also, but at the
m0_rpc__post_locked() which is used to send requests
only - the problem just disappear.

Change-Id: Ic1da7a8281ac259db6e24cf0b8790bea9da0e191