m0t1fs/st: don't change working directory for ST process
The idea is to change working directory only for child shell processes.
ST scripts use `dirname` and this leads to a problem when m0t1fs_test.sh
is executed with relative path instead of absolute. In particular, this
patch fixes issue when ST can't load m0loop.ko:

  insmod ./m0t1fs/linux_kernel/st/../../../mero/m0loop.ko
  insmod: can't read './m0t1fs/linux_kernel/st/../../../mero/m0loop.ko':
          No such file or directory

because working directory was changed to /var/mero/systest-2266/ios4

Change-Id: I47cb11045cc85eaeddfb94d5730ec381865bf965