build: fix `make rpms` to work on VMware's hgfs
VMware uses 'hgfs' file system to share files from host OS with VM, but
this file system doesn't allow to change file's owner, but instead all
files on hgfs have the same owner as on host OS. This can make
`rpmbuild` go nuts if VM doesn't have the same user-id registered in the
system.

With this patch we move dist archive in ~/rpmbuild/SOURCES and change
it's owner to a current user/group.

Change-Id: If5af67267c351bf9cc2ababec6d2607098f734fd