scripts/check-public-api: don't treat undefined functions as warning
Only fail the check if there are some functions present in libmero.so
which are not listed in 'mero/mero-pub.api' file. Previous version of
the script would produce the warning in the reverse case, i.e. some func
is listed in public API file, but missing in libmero.so. But due to we
have some functions that can be public in debug mode, but private in
release mode, it's not possible easily distinguish them from leftovers
in mero-pub.api. So to keep things simple (format of mero-pub.api) we
just don't count such cases as an error.

Change-Id: I0984ac2405e0a2e6c0cdf75cafe9f0dbd8090f6b