net/lnet: MADV_DONTFORK user-kernel shared memory
The memory allocated with nlx_core_mem_alloc()
is used for kernel-user-space sharing. For example,
like ctm_bevq (events queue). So let's advise the
kernel to avoid COW on it on some fork(2)s.

This change fixes the MERO-362 issue:
"rm-ut results in kernel panic". It also could be
the root cause of long running MERO-368 issue:
"Unexpected page migration".

mlock(2) this memory also to be on the safe side.

Reviewed-by: Maxim Medved <max_medved@xyratex.com>
Reviewed-by: Nikita Danilov <nikita_danilov@xyratex.com>

Change-Id: I286e99f7cbfa784ed430738a4ffacf737c6dc6a3