module: simplify finalisation by counting dependencies in use
New field added --- m0_module::m_level_nrefs. Each element of this array
is equal to the number of active dependencies of the corresponding level.
When m0_module__fini() downgrades a level, it tries to perform complete
finalisation of the modules, which this level depends on (i.e., it
attempts to downgrade the level of those modules to M0_MODLEV_NONE).

The previous version of m0_module_fini() did "weak" finalisation of
module dependencies, resulting in convoluted and user-unfriendly
finalisation procedure.

Change-Id: I7ae2ae1df9cf14fe67735cd6c6d940978711f9a8
Reference: https://trello.com/c/tiQ8rLq7/50-init-cleanup-fini