be/tx_group_fom: race fixed at fom wakeup
be_tx_group_fom_iff_waiting_wakeup() may try to
wake up the fom twice. Though it checks that
the fom not waken up already, but it does not work,
because m0_fom_wakeup() does not wake up it immediately,
but just schedules an AST. So we may have the situation
when this AST is tried to be scheduled twice. That
causes the following panic:

"(ast->sa_next == ((void *)0)) at m0_sm_ast_post()"

The fix is pretty straightforward. Instead of calling
m0_fom_wakeup() we can directly call m0_fom_ready()
because we are already in the same locality where
the fom is running.

This addresses MERO-397 issue.

Reviewed-by: Maxim Medved <max_medved@xyratex.com>

Change-Id: I6e25f751502ebede92aed78c99caf00c40b0785c