ut: use plain fixed size array to store tests in suite structure
This is necessary to prevent symbol relocation issue if libmero-ut is
linked as a dynamic library (i.e. .so). If flexible array is used inside
m0_ut_suite structure to store m0_ut structures, then dynamic loader
doesn't relocate it, which renders address of 'ts_tests' to be invalid
in runtime which in turn leads to situation when no tests are executed
in each suite, because 'ts_tests' happens to point to a zeroed memory.

Change-Id: If4fc86fc86027829f6fe50c10e0c59bc32c7b7f2