rpc/formation: item_move_to_urgent_queue() fixed
item_move_to_urgent_queue() should not call frm_which_qtype()
and check whether it is the correct time to move or not
the item to the urgent queue - it is hidden logic which
leads to hidden and hard reproduce bugs, such as MERO-428,
where the item could stuck in the waiting queue.

This addresses MERO-428 issue.

Change-Id: If39c67e9f554766b5338ad597d023f8c517b9422