MERO-232: m0t1fs: error path at ioreq_iosm_handle() fixed
In ioreq_iosm_handle(), when the request after acquiring the lock goes
to IRS_{READING|WRITING} state and fails, it moves to IRS_FAILED state.
After this it will panic as there is no allowed transition to unlock
state from failed state.

To resolve this, unnecessary IO lock acquire and relinquish states are
removed.

Change-Id: I7c03d8a3f244eec265665b00eeab022d7300ba0d