rpc: fix excess reference put of reply item in case of sending error
->ri_reply is set for a request item before sending its reply item
in m0_rpc_item_send_reply(). An extra reference is acquired for this
reply item in sending. This extra reference will be put along with request
put in m0_fop_put() in m0_fom_fini(). But if the reply item is failed
to send, no extra reference is acquired for that reply item. So this
->ri_reply must be cleared in this error case.

Also:
1. Added test case to cover rpc item reply error code path.
2. Added more assertion code in m0_rpc_item_fini() to catch bug.
3. Added timeout value to client session & connection termination.
4. Verify rpcmachine lock is held in various rpc item queue/list operation
   to eliminate race conditions.

https://jira.xyratex.com/browse/MERO-430
Reviewed-by: Andriy Tkachuk
Reviewed-by: Nikita Danilov

Change-Id: Ia374a5fa1bb6c31b007b98deb34b76974aa9157d