m0mkfs: eliminate dependency on on m0mero.ko module
m0mkfs by itself doesn't require any functionality of m0mero.ko module,
it had to use a helper script (m0mkfs.sh) to load it because by default,
libmero.so tries to read "addb node uuid" from loaded m0mero.ko module.
As this uuid is not necessary for correct m0mkfs operation, we
initialize libmero.so in such a way that it doesn't check this uuid, but
uses a null uuid instead.

Change-Id: I1bd1446bd514bbed9a53ff84beca408308f44963