m0t1fs/dir.c: fix race at cobs operations
The problem was introduced at commit 09c0a46:
"improve files creation/deletion speed".
The race was between sending fops and handling
replies in different threads with use of the
common counter and the semaphore. We increased
the counter on sending and decreased on receiving
replies in different threads. When all fops were
sent we down() the semaphore waiting for all the
replies. In replies handling thread when the
counter dropped to zero semaphore was up()-ed and
the waiting sending thread was waken up.

The race here is following: it is possible that
all replies could be received for all the requests
which are sent to the moment, but before the total
number of requests are sent yet. In this case the
counter will become zero and the semaphore will be
up()-ed too early. Thus, down() won't wait for the
last fops replies causing data corruption and
eventual kernel panic.

The solution we chose is simple: drop the counter
and use the semaphore only: down() it in sending
thread as much as fops were send, and up() it
on every reply. It is not very optimal, since it
may cause context switches for every fop reply
in th worst case, but it is simple and reliable.

Reviewed-by: Hua Huang <hua.huang@seagate.com>

Change-Id: I56774b23d4c41c96058af8e8809cfdd0fcc31dc8