MERO-701: assertion failure through m0_stob_io_launch while running SNS rebalance
From the BE documentation in be/tx.h
	- serializibility and concurrency concerns are delegated to a user.
	  Specifically, the user must call m0_be_tx_capture() while the lock,
	  protecting the memory being captured is held. In the current
	  implementation this lock must be held at least until the transaction
	  is closed, in the future this requirement will be weakened;

Concurrent access to the same ad stob with respect to reading/writing needs to be
protected as per above documentation.
SNS repair manages this by assigning locality to copy packet foms based on the
cob_fid, this works as sns repair uses accumulator copy packets to write data.
Thus separate copy packet foms are used for reading and writing.
SNS rebalance unlike SNS repair uses same copy packet (used for reading) fom, by
changing the target cob fid for writing, this leads to the above concurrency concern
as the earlier reqh locality was assigned based on the different cobfid (used for
reading) and now the writing happens on different cob with the same copy packet fom.

This patch fixes this by creating a duplicate copy packet and finalising the earlier
copy packet. The metadata from earlier copy packet is copied to the new copy packet
and the data buffer is detached from the source packet and attached to the new
copy packet to avoid data copy.

* Enhance sns repair st.
* Allow fctx finalisation from every fctx state.
* Release fctx reference in ag_next if that fid is already processed through
  sns data iterator (sns/cm/iter.[ch]).

Reviewed by: Hua Huang <hua.huang@seagate.com>

Change-Id: Iba3e6958b809e2e72204fa2181275d2de535d539